Turns a tour path array into a long data frame.

Internal function, many end users will not need this. Takes the result of manual_tour() or tourr::save_history(). Restructures the array of interpolated bases into a long data frame for use in ggplots.

array2df(
  basis_array,
  data = NULL,
  basis_label = NULL,
  data_label = rownames(data),
  do_center_frame = TRUE
)

Arguments

basis_array

A full (p, d, n_frames) interpolated basis array of a tour, the output of manual_tour or save_history(*_tour()).

data

Optional, (n, p) dataset to project, consisting of numeric variables.

basis_label

Labels for basis display, a character vector with length equal to the number of variables. Defaults to NULL; 3 character abbreviation from colnames of data or rownames of basis.

data_label

Labels for plotly tooltip display. Defaults to the rownames of data. If null, initializes to 1:nrow(data).

do_center_frame

Whether or not to center the mean within each animation frame. Defaults to TRUE.

Examples

## !!This function is not meant for external use!!
dat_std <- scale_sd(wine[, 2:6])
clas <- wine$Type
bas <- basis_pca(dat_std)
mv <- manip_var_of(bas)

## Radial tour array to long df, as used in play_manual_tour()
mt_array <- manual_tour(basis = bas, manip_var = mv)
ls_df_frames <- array2df(basis_array = mt_array, data = dat_std,
                         basis_label = paste0("MyLabs", 1:nrow(bas)))
str(ls_df_frames)
#> List of 2
#>  $ basis_frames:'data.frame':	5 obs. of  4 variables:
#>   ..$ x      : num [1:5] 0.0934 0.4287 0.6507 0.5784 0.2223
#>   ..$ y      : num [1:5] 0.658 -0.134 0.184 -0.434 0.572
#>   ..$ frame  : num [1:5] 1 1 1 1 1
#>   ..$ tooltip: chr [1:5] "MyLabs1" "MyLabs2" "MyLabs3" "MyLabs4" ...
#>   ..- attr(*, "manip_var")= Named int 4
#>   .. ..- attr(*, "names")= chr "Alcohol"
#>  $ data_frames :'data.frame':	178 obs. of  4 variables:
#>   ..$ x      : num [1:178] 0.541 0.306 0.638 0.573 0.788 ...
#>   ..$ y      : num [1:178] 0.976 0.684 0.558 0.866 0.66 ...
#>   ..$ frame  : num [1:178] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ tooltip: int [1:178] 1 2 3 4 5 6 7 8 9 10 ...

## tourr::save_history tour array to long df, as used in play_tour_path()
gt_array <- tourr::save_history(data = dat_std, max_bases = 10)
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.644698 
#> target_dist - cur_dist: 1.644698 
#> generation:  dist =   1.800256 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.518885 
#> target_dist - cur_dist: 0 
#> generation:  dist =   0.7548014 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.643168 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.433034 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.507355 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.736113 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.472626 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.646717 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.83651 
ls_df_frames2 <- array2df(basis_array = gt_array, data = dat_std)
str(ls_df_frames2)
#> List of 2
#>  $ basis_frames:'data.frame':	50 obs. of  4 variables:
#>   ..$ x      : num [1:50] 0.362 -0.192 0.661 -0.441 -0.447 ...
#>   ..$ y      : num [1:50] 0.1458 -0.4362 -0.6143 -0.6405 0.0292 ...
#>   ..$ frame  : num [1:50] 1 1 1 1 1 2 2 2 2 2 ...
#>   ..$ tooltip: chr [1:50] "Alch" "Mlc" "Ash" "Alcl" ...
#>  $ data_frames :'data.frame':	1780 obs. of  4 variables:
#>   ..$ x      : num [1:1780] 0.554 0.584 0.602 0.61 0.558 ...
#>   ..$ y      : num [1:1780] 0.623 0.766 0.436 0.567 0.325 ...
#>   ..$ frame  : num [1:1780] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ tooltip: int [1:1780] 1 2 3 4 5 6 7 8 9 10 ...