Solve for the last basis of a guided tour.

Performs simulated annealing on the index function, solving for it's local extrema. Returns only the last identified basis of the optimization. A truncated, muted extension of tourr::save_history(guided_tour())).

basis_guided(data, index_f = tourr::holes(), d = 2, ...)

Arguments

data

Numeric matrix or data.frame of the observations.

index_f

The index function to optimize. {tourr} exports holes(), cmass(), and lda_pp(class).

d

Number of dimensions in the projection space.

...

Optional, other arguments to pass to tourr::guided_tour

Value

Numeric matrix of the last basis of a guided tour.

See also

Examples

dat_std <- scale_sd(wine[, 2:6])
basis_guided(data = dat_std, index_f = tourr::holes())
#>                 [,1]       [,2]
#> Alcohol    0.5665057  0.3005831
#> Malic      0.3043681 -0.3421526
#> Ash        0.5437994  0.1609096
#> Alcalinity 0.4701990  0.1566823
#> Magnesium  0.2638683 -0.8614756

basis_guided(data = dat_std, index_f = tourr::cmass(),
             alpha = .4, cooling = .9, max.tries = 10, n_sample = 4)
#>                  [,1]       [,2]
#> Alcohol     0.5491221 -0.1203398
#> Malic      -0.3407678  0.1807168
#> Ash        -0.4983716 -0.6026993
#> Alcalinity  0.4601108  0.1900371
#> Magnesium  -0.3496656  0.7439753