Performs a rotation on the manipulation space of the given manip var.

A specific R3 rotation of the manipulation space for a 2D tour. Typically called by manual_tour(). The first 2 columns are x and y in the projection plane. The 3rd column extends "in the z-direction" orthogonal to the projection plane.

rotate_manip_space(manip_space, theta, phi)

Arguments

manip_space

A (p, d+1) dim matrix (manipulation space) to be rotated.

theta

Angle (radians) of "in-projection-plane" rotation (ie. on xy- of the projection). Typically set by the manip_type argument in proj_data().

phi

Angle (radians) of "out-of-projection-plane" rotation (ie. into the z-direction of the manipulation space. Effectively changes the norm of the manip_var in the projection plane.

Value

A (p, d+1) orthonormal matrix of the rotated (manipulation) space. The first 2 columns are x and y in the projection plane. The 3rd column extends "in the z-direction" orthogonal to the projection plane.

See also

Examples

## Setup
dat_std <- scale_sd(wine[, 2:6])
bas <- basis_pca(dat_std)
mv <- manip_var_of(bas)
msp <- create_manip_space(basis = bas, manip_var = mv)
rotate_manip_space(msp, theta = runif(1, max = 2 * pi),
                   phi = runif(1, max = 2 * pi))
#>                   PC1        PC2   manip_sp
#> Alcohol    0.02600859  0.5260390 -0.5260766
#> Malic      0.20721037 -0.5670776 -0.1831696
#> Ash        0.30371880 -0.4946836 -0.5519520
#> Alcalinity 0.92526336  0.2443841  0.2901105
#> Magnesium  0.08961036  0.3118823 -0.5485228

## d = 1 case 
bas1d <- basis_pca(dat_std, d = 1)
mv <- manip_var_of(bas1d)
msp <- create_manip_space(bas1d, mv)
rotate_manip_space(msp, theta = 0, phi = runif(1, max = 2 * pi))
#>                   PC1    manip_sp
#> Alcohol    0.08797566 -0.08595326
#> Malic      0.40387109 -0.39458687
#> Ash        0.69883752  0.71528045
#> Alcalinity 0.54489474 -0.53236866
#> Magnesium  0.20943797 -0.20462339