Returns the axis scale and position.

Internal function. Typically called by other functions to scale the position of the axes data.frame or another data.frame to plot relative to the data.

map_relative(
  x,
  position = c("center", "left", "right", "bottomleft", "topright", "off", "top1d",
    "floor1d", "bottom1d", "full", "facetleft", "facetright", "facettop", "facetbottom"),
  to = NULL
)

Arguments

x

Numeric matrix or data.frame, first 2 columns and scaled and offset the to object.

position

Text specifying the position the axes should go to. Defaults to "center" expects one of: c("center", "left", "right", "bottomleft", "topright", "off", "full", "top1d", "floor1d", "bottom1d", "full", "facetleft", "facetright", "facettop", "facetbottom").

to

Data.frame to scale to. Based on the min/max of the first 2 columns. If left NULL defaults to data.frame(x = c(0, 1), y = c(0, 1).

Value

Transformed values of x, dimension and class unchanged.

See also

Examples

## !!This function is not meant for external use!!
rb <- tourr::basis_random(4, 2)

map_relative(x = rb, position = "bottomleft")
#>           [,1]        [,2]
#> [1,] 0.2988756  0.01007379
#> [2,] 0.3043471 -0.14917209
#> [3,] 0.4145758 -0.12156145
#> [4,] 0.4234139  0.15927656
map_relative(x = rb, position = "right", to = wine[, 2:3])
#>          [,1]     [,2]
#> [1,] 16.38287 3.315936
#> [2,] 16.40782 2.589775
#> [3,] 16.91047 2.715680
#> [4,] 16.95077 3.996301